(De)Serialize difficulty as uint

+++ shared/Converter/DifficultyConverter.cs
@@ -11,7 +11,7 @@ public class DifficultyConverter : JsonConverter<Difficulty>
     {
         return new Difficulty
         {
-            Value = (uint)reader.GetInt32()
+            Value = reader.GetUInt32()
         };
     }
